+++ src/services/dexes/uniswapV3/UniswapV3MultiFetcher.ts
@@ -43,8 +43,9 @@ class UniswapV3MultiFetcher {
   @inject(UniswapV3PoolRepository) protected uniswapV3PoolRepository!: UniswapV3PoolRepository;
   @inject(ContractAddressService) contractAddressService!: ContractAddressService;
 
-  async apply(inputs: UniswapV3MultiFetcherParams[]): Promise<OutputValues[]> {
+  async apply(inputs: UniswapV3MultiFetcherParams[]): Promise<(OutputValues | undefined)[]> {
     this.logger.debug(`[UniswapV3MultiFetcher]: start with inputs ${JSON.stringify(inputs)}`);
+
     if (inputs.length === 0) {
       this.logger.debug('[UniswapV3MultiFetcher] no inputs to fetch');
       return [];
@@ -103,7 +104,7 @@ class UniswapV3MultiFetcher {
   private async fetchData(
     chainId: ChainsIds,
     poolsToFetch: {pool: string; base: string; quote: string; amountInDecimals: number}[],
-  ): Promise<OutputValues[]> {
+  ): Promise<(OutputValues | undefined)[]> {
     const abi = JSON.parse(readFileSync(__dirname + '/UniswapV3FetcherHelper.abi.json', 'utf-8')).abi as never;
     const contract = await this.contractAddressService.getContract(chainId, 'UniswapV3FetcherHelper', abi);
 
@@ -127,30 +128,25 @@ class UniswapV3MultiFetcher {
   private processResult(
     results: {success: boolean; price: BigNumber}[],
     poolsToFetch: {pool: string; base: string; quote: string; amountInDecimals: number}[],
-  ): OutputValues[] {
-    const outputs: OutputValues[] = [];
+  ): (OutputValues | undefined)[] {
+    const outputs: (OutputValues | undefined)[] = [];
 
     for (const [i, result] of results.entries()) {
-      if (!result.success) {
-        outputs.push({base: poolsToFetch[i].base, quote: poolsToFetch[i].quote, value: '0'});
+      const {base, quote} = poolsToFetch[i];
 
-        this.logger.debug(
-          `[UniswapV3MultiFetcher] failed to fetch: ${poolsToFetch[i].base}, ${poolsToFetch[i].quote}: 0`,
-        );
+      if (!result.success) {
+        outputs.push(undefined);
+        this.logger.error(`[UniswapV3MultiFetcher] failed to fetch: ${base}, ${quote}`);
         continue;
       }
 
-      const value = result.price.toString();
-
       outputs.push({
-        base: poolsToFetch[i].base,
-        quote: poolsToFetch[i].quote,
-        value: ethers.utils.formatUnits(ethers.utils.parseUnits(value, 18), 18).toString(),
+        base,
+        quote,
+        value: ethers.utils.formatUnits(result.price.toString(), 18),
       });
 
-      this.logger.debug(
-        `[UniswapV3MultiFetcher] resolved price: ${poolsToFetch[i].base}, ${poolsToFetch[i].quote}: ${result.price}`,
-      );
+      this.logger.debug(`[UniswapV3MultiFetcher] resolved price: ${base}, ${quote}: ${result.price.toString()}`);
     }
 
     return outputs;